lead_gen_agent/requirements.txt
# Lead Generation Agent - Requirements
# Python 3.10+ required

# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.6.0

# Web framework
streamlit>=1.30.0
plotly>=5.18.0

# HTTP & Parsing
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data handling
pydantic>=2.5.0
pandas>=2.0.0

# CLI
click>=8.1.0

# Storage (optional)
sqlalchemy>=2.0.0

# Development
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
